Refactor and extend scoped objects
There is now an array of object classes in multi_site_system which lists all the types of objects which should be treated as site-specific.  Classes in this array will always have an associated site

The approach to scoping searches has been reworked.  The original approach was to use named scopes on each object to select either Object.by_site or Object.by_site_with_descendants.  These scopes were invoked in each controller.  This refactor eliminates the need to explicitly define the scope for each controller by wrapping Spree::BaseController in a with_scope call which scopes each class defined above to the request's site.  This approach eliminates redundancy in the controllers and reduces the potential for accidentally leaving classes un-scoped.  Note that one by-product of this is that it is no longer possible to modify objects which do not belong to the request's scope (without monkeying around with the controller) - this is an intended behavior, as it reduces the probability of users in separate sites modifying each other's data.

The behavior of scoped classes has been changed slightly.  The with_scope function defaults to Object.by_site_with_descendants (rather than Object.by_site, which was the default for the Products controller).  Replicating the original behavior can be implemented by reverting the controller base_scope to .by_site.

The objective of this refactor was to make it easier to scope new classes and to simplify the controller code required.  Hopefully the with_scope approach will make the extension more robust as the Spree codebase evolves.

Adding tests for taxonomies and a .by_site method to satisfy the tests

adding product groups

adding productGroup to scoped classes, refactored test and class_eval code to be less redundant

In the process of refactoring how product scope is defined.

Haven't switched to new approach yet, but the tests are passing (had to change some behavior in the old code to match how *I* want them to behave)

Refactor complete